labs:lab01:ex01: Fix default construction
* Make sure to properly add default values for person class
* Small fix on test

+++ labs/lab01/ex01/src/person.cpp
@@ -4,7 +4,7 @@ void Person::print() const
 {
     std::cout << "========== Person Details ==========" << std::endl;
     std::cout << "Nome: " << name() << std::endl;
-    std::cout << "Idade: " << age() << " anos" << std::endl;
+    std::cout << "Idade: " << static_cast<int>(age()) << " anos" << std::endl;
     std::cout << "Altura: " << height() << " m" << std::endl;
     std::cout << "=====================================" << std::endl;
 }
